feat(list): upgrade list-lite to list
BREAKING CHANGE: Refactor list component with new api.

fix #754
