+++ sonar-project.properties
@@ -1,5 +1,5 @@
 sonar.organization=growingio
-sonar.projectKey=growingio_gio-design
+sonar.projectKey=gio-design
 sonar.sources=src
 
 # lint
@@ -10,5 +10,5 @@ sonar.eslint.reportPaths=eslint-report.json
 sonar.javascript.exclusions=**/__tests__/data.ts
 
 # test coverage
-sonar.coverage.exclusions=**/*test*/*,**/style/*,**/index.*,**/interfaces.*,**/*.stories.*,**/list/*
+sonar.coverage.exclusions=**/*test*/*,**/style/*,**/index.*,**/interfaces.*,**/*.stories.*
 sonar.javascript.lcov.reportPaths=coverage/lcov.info
