--- src/components/list-lite/style/index.ts
@@ -1,2 +1,0 @@
-import '../../empty/style';
-import './index.less';
