+++ src/components/list/index.tsx
@@ -1,3 +1,5 @@
 import List from './List';
 
+export { ListProps } from './interfaces';
+
 export default List;
