--- src/components/list/style/base.less
@@ -1,83 +1,0 @@
-@import '../../../stylesheet/index.less';
-
-@list-prefix-cls: ~'@{component-prefix}-list';
-@select-prefix-cls: ~'@{component-prefix}-select';
-@sortable-cls: ~'@{component-prefix}-sortable';
-
-@link-font-color-normal: #313e75;
-@link-font-color-active: #1248e9;
-@link-font-color-disable: #a2adc8;
-@link-font-color-title: #1248e9;
-@link-font-color-help-text: #1248e9;
-
-@link-font-normal-witght: 14px;
-@link-normal-active: 14px;
-
-@color-background-list-single-hover: #ebedf5; // $palette-grey-3
-@color-background-list-single-normal: #fff; // $palette-white
-@color-background-list-single-click: #f1f2f8; // $palette-grey-2
-@color-background-list-single-active: #f7f8fc; // $palette-grey-1
-@color-background-list-single-disable: #fff; // $palette-white
-@color-background-list-multiple-normal: #fff; // $palette-white
-@color-background-list-multiple-hover: #ebedf5; // $palette-grey-3
-@color-background-list-multiple-click: #f1f2f8; // $palette-grey-2
-@color-background-list-multiple-active: #fff; // $palette-white
-@color-background-list-multiple-disable: #fff; // $palette-white
-
-@color-font-list-primary-normal: #313e75; // $color-black-6
-@color-font-list-primary-hover: #313e75; // $color-black-6
-@color-font-list-primary-active: #1248e9; // $color-blue-5
-@color-font-list-primary-disable: #c7cbd8; // $color-black-2
-@color-font-list-secondary-normal: #adb2c2; // $color-black-3
-@color-font-list-secondary-hover: #adb2c2; // $color-black-3
-@color-font-list-secondary-active: #1248e9; // $color-blue-5
-@color-font-list-secondary-disable: #c7cbd8; // $color-black-2
-@color-font-list-title-normal: #1248e9; // $color-blue-5
-
-.@{list-prefix-cls}-wrapper {
-  display: inline-block;
-  box-sizing: border-box;
-  // background-color: @color-background-list-single-normal;
-  // &:focus {
-  //   outline: none;
-  // }
-  padding: 16px 8px;
-  font-weight: @weight-font-regular;
-
-  &-item {
-    box-sizing: border-box;
-    color: @link-font-color-normal;
-    background-color: @color-background-list-single-normal;
-    user-select: none;
-
-    :hover {
-      background-color: @color-background-list-single-hover;
-      border-radius: 4px;
-      cursor: pointer;
-    }
-
-    :active {
-      background-color: @color-background-list-single-active;
-    }
-  }
-
-  .@{select-prefix-cls}-core {
-    position: relative;
-
-    .@{select-prefix-cls}-list-wrapper {
-      background-color: #fff;
-
-      .@{select-prefix-cls}-list {
-        &:focus {
-          outline: none;
-        }
-      }
-    }
-  }
-
-  &-title {
-    display: inline-block;
-    height: 40px;
-    padding: 8px 15px;
-  }
-}
