--- src/components/list/style/option.less
@@ -1,109 +1,0 @@
-@import './base.less';
-
-.@{select-prefix-cls}-option {
-  box-sizing: border-box;
-  max-width: 100%;
-  padding: 0 25px 0 16px;
-  overflow: hidden;
-  color: @color-font-list-primary-normal;
-  font-size: 14px;
-  line-height: 44px;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  border-radius: 5px;
-  transition: background-color 0.3s ease;
-
-  &:not(.group) {
-    display: flex;
-    align-items: center;
-    cursor: pointer;
-
-    &:hover {
-      background-color: @color-background-list-single-hover;
-    }
-  }
-
-  &.selected {
-    color: @color-font-list-primary-active;
-    background-color: @color-background-list-single-active;
-
-    &:hover {
-      background-color: @color-background-list-single-hover;
-    }
-
-    .@{list-prefix-cls}-option-title {
-      color: @color-font-list-primary-active;
-    }
-
-    .@{list-prefix-cls}-option-desc {
-      color: @color-font-list-secondary-active;
-    }
-  }
-
-  &.multiple {
-    &.selected {
-      color: @color-font-list-primary-normal;
-      background-color: inherit;
-
-      .@{list-prefix-cls}-option-title {
-        color: @color-font-list-primary-normal;
-      }
-
-      .@{list-prefix-cls}-option-desc {
-        color: @color-font-list-secondary-normal;
-      }
-    }
-
-    &:hover {
-      background-color: @color-background-list-multiple-hover;
-    }
-  }
-
-  .@{list-prefix-cls}-option-title {
-    color: @color-font-list-primary-normal;
-  }
-
-  .@{list-prefix-cls}-option-desc {
-    color: @color-font-list-secondary-normal;
-  }
-
-  &.disabled {
-    color: @color-font-list-primary-disable;
-    background-color: transparent !important;
-    cursor: not-allowed;
-
-    .@{list-prefix-cls}-option-title {
-      color: @color-font-list-primary-disable;
-    }
-
-    .@{list-prefix-cls}-option-desc {
-      color: @color-font-list-secondary-disable;
-    }
-  }
-
-  &.group {
-    padding: 0 15px 0 15px;
-    font-size: 12px;
-    line-height: 20.5px;
-  }
-
-  &.group:not(:nth-child(1)) {
-    border-top: 0;
-    border-top-left-radius: 0;
-    border-top-right-radius: 0;
-  }
-
-  &.group:not(:nth-child(1))::before {
-    position: absolute;
-    top: 4px;
-    left: 0;
-    width: 100%;
-    height: 1px;
-    background: #ebedf5;
-    content: '';
-  }
-
-  &.indented {
-    padding-left: 43px;
-  }
-}
