--- src/components/list/style/sort.less
@@ -1,146 +1,0 @@
-@import './base.less';
-
-@icon-circle-color-light: #f1f1fa;
-@icon-circle-color-text: #220968;
-
-.pa_sider-selected-item {
-  position: relative;
-  display: flex;
-  align-items: center;
-  height: 40px;
-  padding-left: 20px;
-  color: @color-font-list-primary-normal;
-  font-size: 14px;
-  line-height: 40px;
-  background: white;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-
-  &.collapsed {
-    display: block;
-    padding: 0;
-  }
-
-  &__name-wrap {
-    position: relative;
-    display: inline-block;
-    height: 40px;
-    padding-left: 4px;
-  }
-
-  &__name {
-    display: inline-block;
-    max-width: 150px;
-    overflow: hidden;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-
-    &.selected {
-      font-size: 14px;
-    }
-  }
-
-  &.indented {
-    padding-left: 30px;
-  }
-
-  &.selected {
-    color: @color-font-list-primary-active;
-    font-weight: bold;
-    background-color: #f2f4fc;
-  }
-
-  &.is-hovered {
-    background-color: #f7f8fc;
-  }
-
-  &:not(.collapsed):not(.is-dragging):not(.is-over).is-hovered {
-    z-index: 1;
-    & .icon-num {
-      display: none;
-    }
-
-    & .icon-drag {
-      display: inline;
-    }
-    & .dh-can-remove {
-      display: inline;
-    }
-  }
-
-  &:hover &__btn-remove {
-    opacity: 1;
-  }
-
-  &__btn-remove {
-    position: absolute;
-    cursor: pointer;
-    opacity: 0;
-
-    &.collapsed {
-      display: none;
-    }
-  }
-
-  .icon-circle {
-    cursor: default;
-  }
-
-  .icon-num {
-    display: inline;
-  }
-
-  .icon-drag {
-    display: none;
-  }
-
-  .dh-can-remove {
-    position: absolute;
-    right: 10px;
-    display: none;
-    width: 12px;
-    height: 40px;
-    &.collapsed {
-      display: none;
-    }
-  }
-}
-
-.pa-sider-icon-circle {
-  position: absolute;
-  left: 2px;
-  display: inline-block;
-  width: 24px;
-  height: 24px;
-  line-height: 23px;
-  text-align: center;
-  cursor: move;
-
-  &.hollow {
-    background-color: #fff;
-    transition: background-color 0.3s ease;
-  }
-
-  &.collapsed {
-    margin: 0;
-    cursor: pointer;
-  }
-
-  &.scaled {
-    width: 32px;
-    height: 32px;
-    font-size: 14px;
-    line-height: 32px;
-    transform: scale(0.6875);
-  }
-}
-
-.@{sortable-cls}-item.is-over {
-  background: #ebebeb;
-  span {
-    display: none;
-  }
-}
-.@{sortable-cls}-item {
-  cursor: move;
-}
