+++ src/components/table/FilterList.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { without, concat } from 'lodash';
-import List from '../list-lite';
+import List from '../list';
 import Checkbox from '../checkbox';
 
 interface FilterListProps {
@@ -8,27 +8,26 @@ interface FilterListProps {
   value: string[];
   onChange: (value: string[]) => void;
   dataSource: {
-    key: string,
-    value: string
+    key: string;
+    value: string;
   }[];
 }
 
 const FilterList = ({ prefixCls, value, onChange, dataSource }: FilterListProps) => (
-    <List className={`${prefixCls}-filter-list`}>
-      {dataSource.map((item) => (
-        <List.Item
-          key={item.key}
-          className={`${prefixCls}-filter-list-item`}
-          onClick={() => {
-            onChange(value.includes(item.key) ? without(value, item.key): concat(value, item.key))
-          }}
-        >
-          <Checkbox checked={value.includes(item.key)} /> 
-          {item.value}
-        </List.Item>
-      ))}
-      
-    </List>
-  )
+  <List className={`${prefixCls}-filter-list`}>
+    {dataSource.map((item) => (
+      <List.Item
+        key={item.key}
+        className={`${prefixCls}-filter-list-item`}
+        onClick={() => {
+          onChange(value.includes(item.key) ? without(value, item.key) : concat(value, item.key));
+        }}
+      >
+        <Checkbox checked={value.includes(item.key)} />
+        {item.value}
+      </List.Item>
+    ))}
+  </List>
+);
 
 export default FilterList;
